+++ flow-typed/libdef.js
@@ -1,0 +1,9 @@
+/*eslint-disable */
+
+declare function cl(a: any): any;
+
+declare var module : {
+  hot : {
+    accept: () => any;
+  };
+};