+++ frontend/js/fn.js
@@ -1,0 +1,5 @@
+/* @flow */
+
+export const add = (a: number, b: number): number => {
+    return a + b;
+};