+++ frontend/js/main.js
@@ -1,3 +1,5 @@
+/* @flow */
+
 import './helpers';
 import svg4everybody from 'svg4everybody';
 import objectFitImages from 'object-fit-images';