+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [1.4.0] - 2017-05-13
+### Added
+- flow
+- babili-webpack-plugin -> minification
+- removing old dependencies
 
 ## [1.3.0] - 2017-03-28
 ### Added